backend/requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.4
django-cors-headers==4.6.0
django-import-export==4.3.3
django-rest==0.8.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
six==1.17.0
sqlparse==0.5.3
tablib==3.7.0
